+++ packages/sdk/src/contracts/index.ts
@@ -1,5 +1,4 @@
 export {
-  encryptedAbi,
   confidentialBalanceOfContract,
   confidentialTransferContract,
   confidentialTransferFromContract,
@@ -16,7 +15,6 @@ export {
 } from "./encrypted";
 
 export {
-  wrapperAbi,
   finalizeUnwrapContract,
   underlyingContract,
   wrapContract,
@@ -27,19 +25,15 @@ export {
   nameContract,
   symbolContract,
   decimalsContract,
+  erc20TotalSupplyContract,
   balanceOfContract,
   allowanceContract,
   approveContract,
 } from "./erc20";
 
-export {
-  deploymentCoordinatorAbi,
-  getWrapperContract,
-  wrapperExistsContract,
-} from "./deployment-coordinator";
+export { getWrapperContract, wrapperExistsContract } from "./deployment-coordinator";
 
 export {
-  erc165Abi,
   ERC7984_INTERFACE_ID,
   ERC7984_WRAPPER_INTERFACE_ID,
   supportsInterfaceContract,
@@ -48,24 +42,32 @@ export {
 } from "./erc165";
 
 export {
-  feeManagerAbi,
   getWrapFeeContract,
   getUnwrapFeeContract,
   getBatchTransferFeeContract,
   getFeeRecipientContract,
 } from "./fee-manager";
 
-export {
-  transferBatcherAbi,
-  confidentialBatchTransferContract,
-  type BatchTransferData,
-} from "./transfer-batcher";
+export { confidentialBatchTransferContract, type BatchTransferData } from "./transfer-batcher";
 
 export {
-  aclAbi,
   delegateForUserDecryptionContract,
   revokeDelegationContract,
   getDelegationExpiryContract,
 } from "./acl";
+
+export {
+  wrappersRegistryAbi,
+  getTokenPairsContract,
+  getTokenPairsLengthContract,
+  getTokenPairsSliceContract,
+  getTokenPairContract,
+  getConfidentialTokenAddressContract,
+  getTokenAddressContract,
+  isConfidentialTokenValidContract,
+  type TokenWrapperPair,
+  type TokenWrapperPairWithMetadata,
+  type PaginatedResult,
+} from "./wrappers-registry";
 
 export { MAX_UINT64 } from "./constants";